Created LocalGameManager and NetworkGameManager to differenciate local game logic of network game logic (wall drop, random value, etc...) + refactoring.

+++ 3 - FBLibrary/GameConfiguration.cs
@@ -84,7 +84,7 @@ namespace FBLibrary
         public static Point BaseTileSize = new Point(32, 32);
 
         // Time
-        public static int DeltaTime; // milliseconds
+        public static long DeltaTime; // Ticks
 
         // TO DELETE
         public const int PlayerNumber = 1;
